chore: move Timing class to core
