+++ packages/flood-runner/src/test-observers/Timing.ts
@@ -6,6 +6,7 @@ import {
 	ResponseTiming,
 	StructuredError,
 } from '@flood/element-api'
+import { Timing } from '@flood/element-core'
 import { NetworkRecordingTestObserver } from './NetworkRecordingTestObserver'
 
 import debugFactory from 'debug'
@@ -17,61 +18,6 @@ const debug = debugFactory('element:grid:timing')
 // 'first-contentful-paint': 'first_contentful_paint',
 // 'first-paint': 'first_paint',
 // }
-
-type TimingRec = { start: number; end: number; thinkTime: number }
-type TimingSegmentName = 'beforeStep' | 'step' | 'afterStep'
-
-class Timing {
-	private segments: Map<TimingSegmentName, TimingRec>
-	constructor(public epoch: Date = new Date()) {
-		this.segments = new Map()
-	}
-
-	start(segmentName: TimingSegmentName) {
-		const now = new Date().valueOf()
-		this.segments.set(segmentName, { start: now, end: 0, thinkTime: 0 })
-	}
-
-	end(segmentName: TimingSegmentName) {
-		const seg = expect(this.segments.get(segmentName), `No timing started for ${segmentName}`)
-		const now = new Date().valueOf()
-		seg.end = now
-		this.segments.set(segmentName, seg)
-	}
-	// TODO thinkTime
-
-	getDurationForSegment(segmentName: TimingSegmentName): number {
-		const { start, end } = expect(
-			this.segments.get(segmentName),
-			`No timing started for ${segmentName}`,
-		)
-		return end - start
-	}
-
-	getThinkTimeForSegment(segmentName: TimingSegmentName): number {
-		const { thinkTime } = expect(
-			this.segments.get(segmentName),
-			`No timing started for ${segmentName}`,
-		)
-		return thinkTime
-	}
-
-	getDurationWithoutThinkTimeForSegment(segmentName: TimingSegmentName): number {
-		return this.getDurationForSegment(segmentName) - this.getThinkTimeForSegment(segmentName)
-	}
-
-	async measureThinkTime(segmentName: TimingSegmentName, func: Function, ...args: any[]) {
-		const seg = expect(this.segments.get(segmentName), `No timing started for ${segmentName}`)
-		const start = new Date()
-		await func.apply(this, ...args)
-		const end = new Date()
-		seg.thinkTime += end.valueOf() - start.valueOf()
-	}
-
-	reset() {
-		this.segments.clear()
-	}
-}
 
 export class TimingObserver extends NetworkRecordingTestObserver {
 	private passed = 0
